+++ opentelemetry-sdk/src/opentelemetry/sdk/_logs/_internal/__init__.py
@@ -78,18 +78,6 @@ class LogDroppedAttributesWarning(UserWarning):
 warnings.simplefilter("once", LogDroppedAttributesWarning)
 
 
-class LogDeprecatedInitWarning(UserWarning):
-    """Custom warning to indicate that deprecated and soon to be deprecated Log classes was used.
-
-    This class is used to filter and handle these specific warnings separately
-    from other warnings, ensuring that they are only shown once without
-    interfering with default user warnings.
-    """
-
-
-warnings.simplefilter("once", LogDeprecatedInitWarning)
-
-
 class LogLimits:
     """This class is based on a SpanLimits class in the Tracing module.
 
@@ -280,6 +268,7 @@ class ReadWriteLogRecord:
             resource=resource,
             instrumentation_scope=instrumentation_scope,
         )
+
 
 class LogRecordProcessor(abc.ABC):
     """Interface to hook the log record emitting action.
@@ -518,7 +507,7 @@ class LoggingHandler(logging.Handler):
                 )
         return attributes
 
-    def _translate(self, record: logging.LogRecord) -> dict:
+    def _translate(self, record: logging.LogRecord) -> LogRecord:
         timestamp = int(record.created * 1e9)
         observered_timestamp = time_ns()
         attributes = self._get_attributes(record)
@@ -552,15 +541,15 @@ class LoggingHandler(logging.Handler):
             "WARN" if record.levelname == "WARNING" else record.levelname
         )
 
-        return {
-            "timestamp": timestamp,
-            "observed_timestamp": observered_timestamp,
-            "context": get_current() or None,
-            "severity_text": level_name,
-            "severity_number": severity_number,
-            "body": body,
-            "attributes": attributes,
-        }
+        return LogRecord(
+            timestamp=timestamp,
+            observed_timestamp=observered_timestamp,
+            context=get_current() or None,
+            severity_text=level_name,
+            severity_number=severity_number,
+            body=body,
+            attributes=attributes,
+        )
 
     def emit(self, record: logging.LogRecord) -> None:
         """
@@ -570,7 +559,7 @@ class LoggingHandler(logging.Handler):
         """
         logger = get_logger(record.name, logger_provider=self._logger_provider)
         if not isinstance(logger, NoOpLogger):
-            logger.emit(**self._translate(record))
+            logger.emit(self._translate(record))
 
     def flush(self) -> None:
         """
@@ -609,6 +598,7 @@ class Logger(APILogger):
     def resource(self):
         return self._resource
 
+    # pylint: disable=arguments-differ
     def emit(self, record: LogRecord):
         """Emits the :class:`ReadWriteLogRecord` by setting instrumentation scope
         and forwarding to the processor.