+++ opentelemetry-sdk/tests/logs/test_logs.py
@@ -17,8 +17,7 @@
 import unittest
 from unittest.mock import Mock, patch
 
-from opentelemetry._logs import LogRecord as APILogRecord
-from opentelemetry._logs import SeverityNumber
+from opentelemetry._logs import LogRecord, SeverityNumber
 from opentelemetry.context import get_current
 from opentelemetry.sdk._logs import (
     Logger,
@@ -178,18 +177,18 @@ class TestLogger(unittest.TestCase):
         self.assertTrue(isinstance(log_record, LogRecord))
         self.assertEqual(log_record.timestamp, None)
         self.assertEqual(log_record.observed_timestamp, 0)
-        self.assertEqual(log_record.context, {})
+        self.assertIsNotNone(log_record.context)
         self.assertEqual(log_record.severity_number, None)
         self.assertEqual(log_record.severity_text, None)
         self.assertEqual(log_record.body, "a log line")
         self.assertEqual(log_record.attributes, {})
         self.assertEqual(log_record.event_name, None)
-        self.assertEqual(log_record.resource, logger.resource)
+        self.assertEqual(log_data.resource, logger.resource)
 
     def test_can_emit_with_keywords_arguments(self):
         logger, log_record_processor_mock = self._get_logger()
 
-        logger.emit(
+        log_record = LogRecord(
             timestamp=100,
             observed_timestamp=101,
             context=get_current(),
@@ -199,16 +198,19 @@ class TestLogger(unittest.TestCase):
             attributes={"some": "attributes"},
             event_name="event_name",
         )
+        logger.emit(log_record)
         log_record_processor_mock.on_emit.assert_called_once()
         log_data = log_record_processor_mock.on_emit.call_args.args[0]
-        log_record = log_data.log_record
-        self.assertTrue(isinstance(log_record, LogRecord))
-        self.assertEqual(log_record.timestamp, 100)
-        self.assertEqual(log_record.observed_timestamp, 101)
-        self.assertEqual(log_record.context, {})
-        self.assertEqual(log_record.severity_number, SeverityNumber.WARN)
-        self.assertEqual(log_record.severity_text, "warn")
-        self.assertEqual(log_record.body, "a body")
-        self.assertEqual(log_record.attributes, {"some": "attributes"})
-        self.assertEqual(log_record.event_name, "event_name")
-        self.assertEqual(log_record.resource, logger.resource)
+        result_log_record = log_data.log_record
+        self.assertTrue(isinstance(result_log_record, LogRecord))
+        self.assertEqual(result_log_record.timestamp, 100)
+        self.assertEqual(result_log_record.observed_timestamp, 101)
+        self.assertIsNotNone(result_log_record.context)
+        self.assertEqual(
+            result_log_record.severity_number, SeverityNumber.WARN
+        )
+        self.assertEqual(result_log_record.severity_text, "warn")
+        self.assertEqual(result_log_record.body, "a body")
+        self.assertEqual(result_log_record.attributes, {"some": "attributes"})
+        self.assertEqual(result_log_record.event_name, "event_name")
+        self.assertEqual(log_data.resource, logger.resource)